Minimal example showing segfault on model flattening over multiple directories (with some debugging).
