+++ src/debug.h
@@ -93,24 +93,26 @@ private:
     bool mNewLine;
 };
 
+void listModelsUnits(const ModelPtr &model);
 void printAnalyserModelEquations(const AnalyserModelPtr &model);
 void printAnalyserModelVariables(const AnalyserModelPtr &model);
 void printAstAsTree(const AnalyserEquationAstPtr &ast);
 void printAstAsCode(const AnalyserEquationAstPtr &ast);
+void printComponent(const ComponentPtr &component, bool includeMaths = true);
 void printComponentMap(const ComponentMap &map);
 void printConnectionMap(const ConnectionMap &map);
 void printEquivalenceMap(const EquivalenceMap &map);
 void printEquivalenceMapWithModelInfo(const EquivalenceMap &map, const ModelPtr &model);
+void printEquivalences(const VariablePtr &variable);
 void printHistory(const History &history);
 void printHistoryEpoch(const HistoryEpochPtr &historyEpoch);
 void printImportLibrary(const ImportLibrary &importlibrary);
+void printModel(const ModelPtr &model, bool includeMaths = true);
+void printNamedPath(const ParentedEntityPtr &parented);
 void printStack(const IndexStack &stack);
 void printStackWithModelInfo(const IndexStack &stack, const ModelPtr &model);
 void printStringStringMap(const StringStringMap &map);
-void printVariableMap(const VariableMap &map);
 void printUnits(const UnitsPtr &units);
-void listModelsUnits(const ModelPtr &model);
-void printNamedPath(const ParentedEntityPtr &parented);
-void printEquivalences(const VariablePtr &variable);
+void printVariableMap(const VariableMap &map);
 
 } // namespace libcellml
