+++ src/importer.cpp
@@ -18,6 +18,7 @@ limitations under the License.
 
 #include <algorithm>
 #include <cmath>
+#include <filesystem>
 #include <fstream>
 #include <libxml/uri.h>
 #include <sstream>
@@ -36,6 +37,8 @@ limitations under the License.
 #include "issue_p.h"
 #include "logger_p.h"
 #include "utilities.h"
+
+#include "debug.h"
 
 namespace libcellml {
 
@@ -329,14 +332,18 @@ std::string pathFromUrl(const std::string &url)
  */
 std::string resolvePath(const std::string &filename, const std::string &base)
 {
-    return pathFromUrl(base) + filename;
+    return std::filesystem::canonical(pathFromUrl(base) + filename).generic_string();
 }
 
 bool Importer::ImporterImpl::fetchModel(const ImportSourcePtr &importSource, const std::string &baseFile)
 {
     std::string url = normaliseDirectorySeparator(importSource->url());
+    Debug() << "Fetch model: " << importSource->url();
+    Debug() << " norm: " << url;
     if (mLibrary.count(url) == 0) {
         url = resolvePath(url, baseFile);
+        Debug() << "Resolved path: " << url;
+        Debug() << "now: " << std::filesystem::canonical(url);
     }
 
     ModelPtr model;
@@ -481,6 +488,7 @@ bool Importer::ImporterImpl::fetchComponent(const ComponentPtr &importComponent,
     }
 
     std::string resolvingUrl = ImporterImpl::resolvingUrl(importComponent->importSource());
+    Debug() << "Component -> resolvingUrl: " << resolvingUrl;
 
     if (encounteredRelatedError) {
         auto issue = Issue::IssueImpl::create();
@@ -572,6 +580,7 @@ bool Importer::ImporterImpl::fetchUnits(const UnitsPtr &importUnits, const std::
     }
 
     std::string resolvingUrl = ImporterImpl::resolvingUrl(importUnits->importSource());
+    Debug() << "Units -> resolvingUrl: " << resolvingUrl;
 
     if (encounteredRelatedError) {
         auto issue = Issue::IssueImpl::create();
@@ -653,6 +662,7 @@ bool Importer::resolveImports(ModelPtr &model, const std::string &basePath)
 
     clearImports(model);
     auto normalisedBasePath = normalisePath(basePath);
+    Debug() << "normalisedBasePath: " << normalisedBasePath;
 
     for (const UnitsPtr &units : getImportedUnits(model)) {
         history.clear();
@@ -670,6 +680,7 @@ bool Importer::resolveImports(ModelPtr &model, const std::string &basePath)
             status = false;
         }
     }
+    printImportLibrary(pFunc()->mLibrary);
 
     return status;
 }
@@ -842,6 +853,8 @@ ComponentPtr flattenComponent(const ComponentEntityPtr &parent, ComponentPtr &co
         }
 
         // Get list of required units from component's variables and math cn elements.
+        printModel(clonedImportModel);
+        printComponent(importedComponentCopy);
         std::vector<UnitsPtr> requiredUnits = unitsUsed(clonedImportModel, importedComponentCopy);
 
         std::vector<UnitsPtr> uniqueRequiredUnits;
