+++ tests/resources/importer/periodicstimulus/components/INa.xml
@@ -1,0 +1,9 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<model name="INa" xmlns="http://www.cellml.org/cellml/1.1#" xmlns:cellml="http://www.cellml.org/cellml/1.1#" xmlns:cmeta="http://www.cellml.org/metadata/1.0#" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <import xlink:href="units.xml">
+        <units name="uA_per_cmsq" units_ref="uA_per_cmsq"/>
+    </import>
+    <component name="INa">
+        <variable name="INa" private_interface="out" public_interface="out" units="uA_per_cmsq"/>
+    </component>
+</model>
