+++ tests/resources/importer/periodicstimulus/components/stimulated.xml
@@ -1,0 +1,23 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<model name="stimulated" xmlns="http://www.cellml.org/cellml/1.1#" xmlns:cellml="http://www.cellml.org/cellml/1.1#" xmlns:cmeta="http://www.cellml.org/metadata/1.0#" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <import xlink:href="units.xml">
+        <units name="uA_per_cmsq" units_ref="uA_per_cmsq"/>
+    </import>
+    <import xlink:href="INa.xml">
+        <component component_ref="INa" name="INa"/>
+    </import>
+    <component name="action_potential">
+<!--        <variable name="Istim" private_interface="out" public_interface="in" units="uA_per_cmsq"/>-->
+        <variable name="INa" private_interface="in" public_interface="out" units="uA_per_cmsq"/>
+    </component>
+    <group>
+        <relationship_ref relationship="encapsulation"/>
+        <component_ref component="action_potential">
+            <component_ref component="INa"/>
+        </component_ref>
+    </group>
+    <connection>
+        <map_components component_1="action_potential" component_2="INa"/>
+        <map_variables variable_1="INa" variable_2="INa"/>
+    </connection>
+</model>
