+++ tests/resources/importer/periodicstimulus/components/units.xml
@@ -1,0 +1,13 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<model name="units" xmlns="http://www.cellml.org/cellml/1.1#" xmlns:cellml="http://www.cellml.org/cellml/1.1#">
+    <units name="cm">
+        <unit prefix="centi" units="metre"/>
+    </units>
+    <units name="uA">
+        <unit prefix="micro" units="ampere"/>
+    </units>
+    <units name="uA_per_cmsq">
+        <unit units="uA"/>
+        <unit exponent="-2" units="cm"/>
+    </units>
+</model>
