+++ tests/resources/importer/periodicstimulus/experiments/periodic-stimulus.xml
@@ -1,0 +1,9 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<model name="periodic_stimulus" xmlns="http://www.cellml.org/cellml/1.1#" xmlns:cellml="http://www.cellml.org/cellml/1.1#" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <import xlink:href="../components/units.xml">
+        <units name="uA_per_cmsq" units_ref="uA_per_cmsq"/>
+    </import>
+    <import xlink:href="../components/stimulated.xml">
+        <component component_ref="action_potential" name="_model"/>
+    </import>
+</model>
