+++ tests/test_utils.cpp
@@ -87,104 +87,6 @@ void printIssues(const libcellml::LoggerPtr &l, bool headings, bool cellmlElemen
     }
 }
 
-static const std::string FIXED_INDENT = "    ";
-
-void printComponent(const libcellml::ComponentPtr &component, size_t c, const std::string &indent, bool includeMaths)
-{
-    if (c == -1) {
-        std::cout << "COMPONENT: '" << component->name() << "'";
-    } else {
-        std::cout << indent << "[" << c + 1 << "]: " << component->name();
-    }
-
-    if (component->id() != "") {
-        std::cout << ", id: " << component->id();
-    }
-
-    std::cout << std::endl;
-    std::cout << indent << FIXED_INDENT << "VARIABLES: " << component->variableCount() << " variables" << std::endl;
-
-    // Printing the variables within the component.
-    for (size_t v = 0; v < component->variableCount(); ++v) {
-        std::cout << indent << FIXED_INDENT << FIXED_INDENT;
-        std::cout << "[" << v + 1 << "]: " << component->variable(v)->name();
-        if (component->variable(v)->units() != nullptr) {
-            std::cout << " [" << component->variable(v)->units()->name() << "]";
-        }
-        if (component->variable(v)->initialValue() != "") {
-            std::cout << ", initial = " << component->variable(v)->initialValue();
-        }
-        std::cout << std::endl;
-        if (component->variable(v)->equivalentVariableCount() > 0) {
-            std::cout << indent << FIXED_INDENT << FIXED_INDENT << FIXED_INDENT;
-            std::string con = "  └──> ";
-            for (size_t e = 0; e < component->variable(v)->equivalentVariableCount(); ++e) {
-                auto ev = component->variable(v)->equivalentVariable(e);
-                if (ev == nullptr) {
-                    std::cout << "WHOOPS! Null equivalent variable!";
-                    continue;
-                }
-                libcellml::ComponentPtr ev_parent = std::dynamic_pointer_cast<libcellml::Component>(ev->parent());
-                if (ev_parent == nullptr) {
-                    std::cout << "WHOOPS! Null parent component for equivalent variable!";
-                    continue;
-                }
-                std::cout << con << ev_parent->name() << ":" << ev->name();
-                if (ev->units() != nullptr) {
-                    std::cout << " [" << ev->units()->name() << "]";
-                }
-                con = ", ";
-            }
-            std::cout << std::endl;
-        }
-    }
-
-    // Print the maths within the component.
-    if (includeMaths) {
-        if (component->math() != "") {
-            std::cout << indent << "  Maths in the component is:" << std::endl;
-            std::cout << component->math() << std::endl;
-        }
-    }
-
-    // Print the encapsulated components
-    if (component->componentCount() > 0) {
-        std::cout << indent << FIXED_INDENT << "CHILD COMPONENTS: " << component->componentCount()
-                  << " child components" << std::endl;
-        std::string newIndent = indent + FIXED_INDENT + FIXED_INDENT;
-
-        for (size_t c2 = 0; c2 < component->componentCount(); ++c2) {
-            auto child = component->component(c2);
-            printComponent(child, c2, newIndent, includeMaths);
-        }
-    }
-}
-
-void printComponent(const libcellml::ComponentPtr &component, bool includeMaths)
-{
-    printComponent(component, -1, {}, includeMaths);
-}
-
-void printModel(const libcellml::ModelPtr &model, bool includeMaths)
-{
-    std::cout << "MODEL: '" << model->name() << "'";
-    if (model->id() != "") {
-        std::cout << ", id: '" << model->id() << "'";
-    }
-    std::cout << std::endl;
-
-    std::cout << FIXED_INDENT << "UNITS: " << model->unitsCount() << " custom units" << std::endl;
-    for (size_t u = 0; u < model->unitsCount(); ++u) {
-        std::cout << FIXED_INDENT << FIXED_INDENT << "[" << u + 1 << "]: " << model->units(u)->name() << std::endl;
-    }
-
-    std::cout << FIXED_INDENT << "COMPONENTS: " << model->componentCount() << " components" << std::endl;
-    for (size_t c = 0; c < model->componentCount(); ++c) {
-        auto component = model->component(c);
-        printComponent(component, c, FIXED_INDENT + FIXED_INDENT, includeMaths);
-    }
-}
-
 template<typename T>
 std::vector<T> expectedValues(size_t size, T value)
 {
